Rename `waitUntilCalled.js` to `wait-until-called.js`
